Move Python SDK message router to module
